Reorganize issue types on subpackages
